[ADD] Add issue new hook point.
